CNN_YOLO_MODEL/requirements.txt
ultralytics==8.0.196
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
scikit-learn==1.3.0
PyYAML==6.0.1
requests==2.31.0
tqdm==4.66.1